Migrate acceptance tests: old Video -> new Video.
