+++ cms/djangoapps/contentstore/features/video-editor.py
@@ -2,18 +2,7 @@
 # pylint: disable=C0111
 
 from lettuce import world, step
-
-
-@step('I see the correct settings and default values$')
-def i_see_the_correct_settings_and_values(step):
-    world.verify_all_setting_entries([['Default Speed', 'OEoXaMPEzfM', False],
-                                      ['Display Name', 'Video', False],
-                                      ['Download Track', '', False],
-                                      ['Download Video', '', False],
-                                      ['Show Captions', 'True', False],
-                                      ['Speed: .75x', '', False],
-                                      ['Speed: 1.25x', '', False],
-                                      ['Speed: 1.5x', '', False]])
+from terrain.steps import reload_the_page
 
 
 @step('I have set "show captions" to (.*)')
@@ -24,9 +13,19 @@ def set_show_captions(step, setting):
     world.css_click('a.save-button')
 
 
-@step('I see the correct videoalpha settings and default values$')
-def correct_videoalpha_settings(_step):
-    world.verify_all_setting_entries([['Display Name', 'Video Alpha', False],
+@step('when I view the (video.*) it (.*) show the captions')
+def shows_captions(_step, video_type, show_captions):
+    # Prevent cookies from overriding course settings
+    world.browser.cookies.delete('hide_captions')
+    if show_captions == 'does not':
+        assert world.css_has_class('.%s' % video_type, 'closed')
+    else:
+        assert world.is_css_not_present('.%s.closed' % video_type)
+
+
+@step('I see the correct video settings and default values$')
+def correct_video_settings(_step):
+    world.verify_all_setting_entries([['Display Name', 'Video', False],
                                       ['Download Track', '', False],
                                       ['Download Video', '', False],
                                       ['End Time', '0', False],
@@ -38,3 +37,12 @@ def correct_videoalpha_settings(_step):
                                       ['Youtube ID for .75x speed', '', False],
                                       ['Youtube ID for 1.25x speed', '', False],
                                       ['Youtube ID for 1.5x speed', '', False]])
+
+
+@step('my video display name change is persisted on save')
+def video_name_persisted(step):
+    world.css_click('a.save-button')
+    reload_the_page(step)
+    world.edit_component()
+    world.verify_setting_entry(world.get_setting_entry('Display Name'), 'Display Name', '3.4', True)
+
